feat: enable basic group crud operations
This commits adds basic group CRUD operations to the REST API:
- POST /api/v1/groups/
- PUT /api/v1/groups/{id}
- DELETE /api/v1/groups/{id}

By default, only the current user is added to a new group. The group is populated with the
built-in plugin parameter types.

The default user group (public) is protected from deletion or modification by any user.

Group deletion removes all members. The constraint for ensuring a user has at least one group is
maintained by preventing deletion of the public group that all users are part of.

This commit adds checks for all resource endpoints to ensure users cannot access, modify, or delete
resources they don't have the correct permissions for. There are not currently protections in place
to prevent users from deleting or renaming groups they are not a member of.

It makes use of the existing UserNotInGroupError when a user attempts to access a resource from a
group they do not belong to, and adds the UserPermissionsError for cases where a user is in the
group but doesn't have the proper permissions to perform an action.

It adds new tests to verify users can not get or delete resources belonging to a group they do not
access to. Modification and permissions checkes remain untested.

+++ src/dioptra/restapi/db/repository/queues.py
@@ -19,10 +19,12 @@ The queue repository: data operations related to queues
 """
 
 from collections.abc import Iterable, Sequence
-from typing import Any, Final, overload
+from typing import Any, Final, cast, overload
+
+from flask_login import current_user
 
 import dioptra.restapi.db.repository.utils as utils
-from dioptra.restapi.db.models import Group, Queue, Resource, Tag
+from dioptra.restapi.db.models import Group, Queue, Resource, Tag, User
 
 
 class QueueRepository:
@@ -124,6 +126,10 @@ class QueueRepository:
         Raises:
             EntityDoesNotExistError: if the queue does not exist
         """
+        user = cast(User, current_user)
+        resource = utils.get_resource(self.session, queue)
+
+        utils.assert_can_delete_resource(self.session, user, resource)
 
         utils.delete_resource(self.session, queue)
